Rename StrahlkorperDatabox to Tags
Tags is the name we've used everywhere else, so this is for consistency.
