+++ src/ApparentHorizons/CMakeLists.txt
@@ -7,8 +7,8 @@ set(LIBRARY_SOURCES
     FastFlow.cpp
     SpherepackIterator.cpp
     Strahlkorper.cpp
-    StrahlkorperDataBox.cpp
     StrahlkorperGr.cpp
+    Tags.cpp
     YlmSpherepack.cpp
     YlmSpherepackHelper.cpp
     )
